+++ moneybox/api/serializers/currency_serializers.py
@@ -1,0 +1,15 @@
+from rest_framework.serializers import ModelSerializer
+
+from wallet.models import Currency, CurrencyRate
+
+
+class CurrencySerializer(ModelSerializer):
+    class Meta:
+        model = Currency
+        fields = '__all__'
+
+
+class CurrencyRateSerializer(ModelSerializer):
+    class Meta:
+        model = CurrencyRate
+        fields = '__all__'
