+++ moneybox/api/serializers/expense_serializers.py
@@ -1,0 +1,15 @@
+from rest_framework.serializers import ModelSerializer
+
+from wallet.models import ExpenseCategory, Expense
+
+
+class ExpenseCategorySerializer(ModelSerializer):
+    class Meta:
+        model = ExpenseCategory
+        fields = '__all__'
+
+
+class ExpenseSerializer(ModelSerializer):
+    class Meta:
+        model = Expense
+        fields = '__all__'
