+++ moneybox/api/views/__init__.py
@@ -1,0 +1,20 @@
+from .currency_views import CurrencyViewSet, CurrencyRateViewSet
+from .expense_views import ExpenseCategoryViewSet, ExpenseViewSet
+from .group_views import GroupViewSet
+from .income_views import IncomeViewSet, IncomeCategoryViewSet
+from .profile_views import ProfileViewSet
+from .transfer_views import TransferViewSet
+from .wallet_views import WalletViewSet
+
+__all__ = [
+    'CurrencyViewSet',
+    'CurrencyRateViewSet',
+    'ExpenseCategoryViewSet',
+    'ExpenseViewSet',
+    'GroupViewSet',
+    'IncomeCategoryViewSet',
+    'IncomeViewSet',
+    'ProfileViewSet',
+    'TransferViewSet',
+    'WalletViewSet',
+]
