+++ moneybox/api/views/currency_views.py
@@ -1,0 +1,14 @@
+from rest_framework.viewsets import ModelViewSet
+
+from api.serializers import CurrencySerializer, CurrencyRateSerializer
+from wallet.models import Currency, CurrencyRate
+
+
+class CurrencyViewSet(ModelViewSet):
+    queryset = Currency.objects.order_by('pk')
+    serializer_class = CurrencySerializer
+
+
+class CurrencyRateViewSet(ModelViewSet):
+    queryset = CurrencyRate.objects.order_by('pk')
+    serializer_class = CurrencyRateSerializer
