+++ moneybox/api/views/group_views.py
@@ -1,0 +1,9 @@
+from rest_framework.viewsets import ModelViewSet
+
+from api.serializers import GroupSerializer
+from wallet.models import Group
+
+
+class GroupViewSet(ModelViewSet):
+    queryset = Group.objects.order_by('pk')
+    serializer_class = GroupSerializer
