+++ moneybox/api/views/income_views.py
@@ -1,0 +1,14 @@
+from rest_framework.viewsets import ModelViewSet
+
+from api.serializers import IncomeCategorySerializer, IncomeSerializer
+from wallet.models import IncomeCategory, Income
+
+
+class IncomeCategoryViewSet(ModelViewSet):
+    queryset = IncomeCategory.objects.order_by('pk')
+    serializer_class = IncomeCategorySerializer
+
+
+class IncomeViewSet(ModelViewSet):
+    queryset = Income.objects.order_by('pk')
+    serializer_class = IncomeSerializer
