+++ moneybox/api/views/profile_views.py
@@ -1,0 +1,9 @@
+from rest_framework.viewsets import ModelViewSet
+
+from api.serializers import ProfileSerializer
+from users.models import Profile
+
+
+class ProfileViewSet(ModelViewSet):
+    queryset = Profile.objects.order_by('pk')
+    serializer_class = ProfileSerializer
