+++ moneybox/api/views/transfer_views.py
@@ -1,0 +1,9 @@
+from rest_framework.viewsets import ModelViewSet
+
+from api.serializers import TransferSerializer
+from wallet.models import Transfer
+
+
+class TransferViewSet(ModelViewSet):
+    queryset = Transfer.objects.order_by('pk')
+    serializer_class = TransferSerializer
