+++ moneybox/api/views/wallet_views.py
@@ -1,0 +1,9 @@
+from rest_framework.viewsets import ModelViewSet
+
+from api.serializers import WalletSerializer
+from wallet.models import Wallet
+
+
+class WalletViewSet(ModelViewSet):
+    queryset = Wallet.objects.order_by('pk')
+    serializer_class = WalletSerializer
