+++ moneybox/moneybox/settings.py
@@ -1,14 +1,3 @@
-"""
-Django settings for moneybox project.
-
-Generated by 'django-admin startproject' using Django 4.2.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/4.2/topics/settings/
-
-For the full list of settings and their values, see
-https://docs.djangoproject.com/en/4.2/ref/settings/
-"""
 import os
 from pathlib import Path
 
@@ -25,7 +14,7 @@ SECRET_KEY = os.getenv("DJANGO_SECRET", "secret")
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = os.getenv("DJANGO_DEBUG", "False") == "True"
 
-ALLOWED_HOSTS = []
+ALLOWED_HOSTS = ['*']
 
 
 # Application definition
@@ -42,6 +31,8 @@ INSTALLED_APPS = [
     "rest_framework.authtoken",
     "drf_generators",
     "api",
+    "wallet",
+    "users",
     "drf_spectacular",
 ]
 
@@ -79,14 +70,21 @@ WSGI_APPLICATION = "moneybox.wsgi.application"
 # Database
 # https://docs.djangoproject.com/en/4.2/ref/settings/#databases
 
+# DATABASES = {
+#     "default": {
+#         "ENGINE": "django.db.backends.postgresql",
+#         "NAME": os.getenv("POSTGRES_DB"),
+#         "USER": os.getenv("POSTGRES_USER"),
+#         "PASSWORD": os.getenv("POSTGRES_PASSWORD"),
+#         "HOST": os.getenv("POSTGRES_DB_HOST", "db"),
+#         "PORT": os.getenv("POSTGRES_DB_PORT", "5342"),
+#     }
+# }
+
 DATABASES = {
-    "default": {
-        "ENGINE": "django.db.backends.postgresql",
-        "NAME": os.getenv("POSTGRES_DB"),
-        "USER": os.getenv("POSTGRES_USER"),
-        "PASSWORD": os.getenv("POSTGRES_PASSWORD"),
-        "HOST": os.getenv("POSTGRES_DB_HOST", "db"),
-        "PORT": os.getenv("POSTGRES_DB_PORT", "5342"),
+    'default': {
+        'ENGINE': 'django.db.backends.sqlite3',
+        'NAME': BASE_DIR / 'db.sqlite3',
     }
 }
