+++ moneybox/moneybox/wsgi.py
@@ -1,12 +1,3 @@
-"""
-WSGI config for moneybox project.
-
-It exposes the WSGI callable as a module-level variable named ``application``.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/4.2/howto/deployment/wsgi/
-"""
-
 import os
 
 from django.core.wsgi import get_wsgi_application
