+++ moneybox/wallet/models.py
@@ -1,8 +1,7 @@
-from django.contrib.auth import get_user_model
 from django.db import models, transaction
 from decimal import Decimal
 
-User = get_user_model()
+from users.models import Profile
 
 
 class TimestampMixin(models.Model):
@@ -72,27 +71,6 @@ class CurrencyRate(TimestampMixin):
     class Meta:
         verbose_name = "Currency rate"
         verbose_name_plural = "Currency rates"
-
-
-class Profile(TimestampMixin):
-    user = models.OneToOneField(
-        User,
-        on_delete=models.CASCADE,
-        verbose_name="User",
-        help_text="User associated with the profile",
-        db_index=True,
-    )
-    first_name = models.CharField(
-        max_length=255, verbose_name="First name", help_text="First name of the user"
-    )
-    last_name = models.CharField(
-        max_length=255, verbose_name="Last name", help_text="Last name of the user"
-    )
-    email = models.EmailField(verbose_name="Email", help_text="Email of the user")
-
-    class Meta:
-        verbose_name = "Profile"
-        verbose_name_plural = "Profiles"
 
 
 class Group(TimestampMixin):
