+++ nginx/default.conf
@@ -21,15 +21,23 @@ http {
         server_name yourdomain.com;  # TODO fix domain
         charset utf-8;
 
+        location /api/ {
+          proxy_pass http://host.docker.internal:8000/api/;
+          proxy_set_header Host $host;
+        }
+
         location /static/ {
             alias /usr/share/nginx/html/;
         }
 
-        location / {
-            proxy_pass http://web:8000/;   # TODO replace it to docker
-            proxy_set_header Host $host;
-            proxy_set_header X-Real-IP $remote_addr;
-            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-        }
+            location / {
+        root /usr/share/nginx/html;
+        index  index.html index.htm;
+        try_files $uri /index.html;
+        proxy_set_header        Host $host;
+        proxy_set_header        X-Real-IP $remote_addr;
+        proxy_set_header        X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header        X-Forwarded-Proto $scheme;
+      }
     }
 }
